pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "music-adder"
version = "0.3.0"
requires-python = ">=3.10"
dependencies = [
    "yt-dlp",
    "pyacoustid",
    "musicbrainzngs",
    "mutagen",
    "rich",
    "pyyaml",
]

[project.scripts]
music-adder = "music_adder.cli:main"
becca-music-adder = "music_adder.cli:becca_main"
